requirements.txt
# core dependencies
Django>=4.2
djangorestframework>=3.14.0
#redis>=4.0
#pymemcache>=4.0.0
#django-redis>=5.0.0
#psycopg2-binary>=2.9.0

# Testing dependencies
pytest>=7.0.0
pytest-django>=4.5.0
pytest-cov>=4.0.0
pytest-xdist>=3.2.0

# build dependencies
twine==6.1.0
build==1.3.0

#documentation dependencies
mkdocs-material==9.1.12
